Rename folders to make them not hidden; add Linux URL to YouTube demo video; update readme
Files and folders that begin with a period (.) on Linux and Mac are
hidden, so I have renamed the following folders to make them not hidden
anymore, while still keeping them near the top of the alphabetical
order, which  seemed to be intentional, and keeping them in the same
order:

1. .Electronics --> 1-Electronics
2. .Mechanics --> 2-Mechanics
3. .Software --> 3-Software

I've also removed spaces in any path to code, as this is best practice
to avoid compiler problems on various systems, and I've moved the
"ArduinoUniqueID" library into a folder by that name.

Lastly, I've cleaned up the readme a bit and added a .desktop file to
the YouTube video. This is a clickable link to open up a browser on a
Linux computer--the same as would happen with a .url file on a Windows
computer. See: "AmboVent/Other/Footage/YouTube AmboVent Demo.desktop".